fix(logging): unify logging init across server CLI and desktop entrypoints
桌面版/嵌入式 server 此前仅用 stdout-only tracing subscriber，不读取
settings.log 也不写 dinotty.log。从 Dock/Finder 启动桌面 app 时
stdout 被 /dev/null 接收，所有日志静默丢失，导致桌面版运行期问题
（如标签页异常消失）无任何磁盘日志可查。

将 src/main.rs 中的日志初始化逻辑下沉为 settings::init_logging()
公共函数，由 src/main.rs 与 src-tauri/src/main.rs 共同调用，使任意
入口启动都按 settings.log.* 统一落盘。

Fixes #143
